// src/payload-types.ts
export interface NewsletterMember {
  id: number;
  name?: string | null;
  provider: 'local' | 'google' | 'microsoft';
  interests?: ('filmes' | 'series' | 'animes' | 'games')[] | null;
  source?: string[] | null;
  subscriptionStatus: 'active' | 'paused';
  lastLoginAt?: string | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'newsletter-members';
}